Rename SNTEventProvider to SNTEndpointSecurityEventHandler
