+++ Source/santad/EventProviders/SNTEndpointSecurityTamperResistance.h
@@ -18,7 +18,7 @@
 
 #include "Source/santad/EventProviders/EndpointSecurity/EndpointSecurityAPI.h"
 #import "Source/santad/EventProviders/SNTEndpointSecurityClient.h"
-#import "Source/santad/EventProviders/SNTEventProvider.h"
+#import "Source/santad/EventProviders/SNTEndpointSecurityEventHandler.h"
 #include "Source/santad/Logs/EndpointSecurity/Logger.h"
 
 @interface SNTEndpointSecurityTamperResistance : SNTEndpointSecurityClient<SNTEventProvider>
